+++ mithril/backends/with_autograd/jax_backend/backend.py
@@ -21,9 +21,11 @@ import jax
 import jax.numpy as jnp
 
 from ....core import Dtype
+from ....cores.jax import ops
+from ....cores.jax import utils as core_utils
 from ...backend import PadWidthType, ParallelBackend
 from ...utils import DtypeSubTypes, StaticScalar, process_shape
-from . import ops, utils
+from . import utils
 from .parallel import JaxParallel
 from .utils import CODEGEN_CONFIG
 
@@ -49,7 +51,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
 
     backend_type = "jax"
     registered_primitives: dict[str, Callable[..., jax.numpy.ndarray]] = {}
-    primitive_fn_path = "mithril.backends.with_autograd.jax_backend.ops"
+    primitive_fn_path = "mithril.cores.jax.ops"
 
     def __init__(
         self,
@@ -59,7 +61,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
         device_mesh: tuple[int, ...] | None = None,
     ) -> None:
         self._device = device
-        utils.get_device(device)  # Check device is available
+        core_utils.get_device(device)  # Check device is available
         self._dtype = dtype
         self._parallel_manager: JaxParallel | None = None
 
@@ -74,7 +76,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
         self.primitive_function_dict = ops.primitive_func_dict
         self.prng_key = jax.random.PRNGKey(self.seed)
 
-        for key, value in utils.dtype_map.items():
+        for key, value in core_utils.dtype_map.items():
             setattr(self, key, value)
 
     @property
@@ -94,7 +96,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
 
     @property
     def device(self) -> jax.Device:
-        return utils.get_device(self._device)
+        return core_utils.get_device(self._device)
 
     def get_device(self) -> Any:
         return self._device
@@ -117,7 +119,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
         list[str]
             List of available devices.
         """
-        return utils.get_available_devices()
+        return core_utils.get_available_devices()
 
     @staticmethod
     def register_primitive(fn: Callable[..., Any]) -> None:
@@ -139,7 +141,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
         device: str
             The target device for the data.
         """
-        _device = utils.get_device(device)
+        _device = core_utils.get_device(device)
         if not asynchronous:
             return jax.device_put(data, device=_device).block_until_ready()
         return jax.device_put(data, device=_device)
@@ -185,7 +187,7 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
         _dtype = utils.determine_dtype(input, dtype, self._dtype, self.precision)
 
         with jax.default_device(self.device):
-            array = jax.numpy.array(input, dtype=utils.dtype_map[_dtype])
+            array = jax.numpy.array(input, dtype=core_utils.dtype_map[_dtype])
 
         if self._parallel_manager is not None:
             array = self._parallel_manager.parallelize(array, device_mesh)
@@ -673,11 +675,11 @@ class JaxBackend(ParallelBackend[jax.numpy.ndarray]):
         default_type: str | None = None,
     ) -> jax.numpy.dtype[Any]:
         if isinstance(dtype, Dtype):
-            return utils.dtype_map[dtype.name]
+            return core_utils.dtype_map[dtype.name]
         elif dtype is None:
             if default_type is None:
                 default_type = self._get_default_subtype()
-            return utils.dtype_map[default_type + str(self.precision)]
+            return core_utils.dtype_map[default_type + str(self.precision)]
         else:
             raise ValueError(f"Invalid dtype {dtype}")
