+++ mithril/backends/with_autograd/mlx_backend/backend.py
@@ -21,10 +21,11 @@ import mlx.core as mx
 import mlx.nn as nn
 
 from ....core import Dtype
+from ....cores.mlx import ops
+from ....cores.mlx import utils as core_utils
 from ...backend import Backend, PadWidthType
 from ...utils import DtypeSubTypes, StaticScalar, process_shape
-from . import ops, utils
-from .utils import CODEGEN_CONFIG
+from . import utils
 
 __all__ = ["MlxBackend"]
 
@@ -35,7 +36,7 @@ class MlxBackend(Backend[mx.array]):
     backend_type = "mlx"
     supported_dtypes = [Dtype.float16, Dtype.bfloat16, Dtype.float32]
     registered_primitives: dict[str, Callable[..., mx.array]] = {}
-    primitive_fn_path = "mithril.backends.with_autograd.mlx_backend.ops"
+    primitive_fn_path = "mithril.cores.mlx.ops"
 
     def __init__(
         self,
@@ -54,7 +55,7 @@ class MlxBackend(Backend[mx.array]):
         self.primitive_function_dict = ops.primitive_func_dict
         self.prng_key = mx.random.key(self.seed)
 
-        for key, value in utils.dtype_map.items():
+        for key, value in core_utils.dtype_map.items():
             setattr(self, key, value)
 
     @property
@@ -75,7 +76,7 @@ class MlxBackend(Backend[mx.array]):
 
     @property
     def codegen_config(self) -> dict[str, bool]:
-        return CODEGEN_CONFIG
+        return utils.CODEGEN_CONFIG
 
     def get_device(self) -> Any:
         return self._device
@@ -192,7 +193,7 @@ class MlxBackend(Backend[mx.array]):
 
     def array(self, input: Any, *, dtype: Dtype | None = None) -> mx.array:
         _dtype = utils.determine_dtype(input, dtype, self._dtype, self.precision)
-        return mx.array(input, dtype=utils.dtype_map[_dtype])
+        return mx.array(input, dtype=core_utils.dtype_map[_dtype])
 
     def zeros(
         self, *shape: int | tuple[int, ...] | list[int], dtype: Dtype | None = None
@@ -622,11 +623,11 @@ class MlxBackend(Backend[mx.array]):
         default_type: str | None = None,
     ) -> mx.Dtype:
         if isinstance(dtype, Dtype):
-            return utils.dtype_map[dtype.name]
+            return core_utils.dtype_map[dtype.name]
         elif dtype is None:
             if default_type is None:
                 default_type = self._get_default_subtype()
-            return utils.dtype_map[default_type + str(self.precision)]
+            return core_utils.dtype_map[default_type + str(self.precision)]
         else:
             raise ValueError(f"Invalid dtype {dtype}")
