+++ mithril/cores/c/compile.sh
