+++ mithril/cores/jax/ops.py
@@ -24,11 +24,8 @@ import jax.scipy.linalg as slin
 from jax import lax, vmap
 from jax import nn as functionals
 
-from .... import core
-from ....utils.type_utils import is_tuple_int
-from ....utils.utils import find_dominant_type
-from ...utils import NestedFloatOrIntOrBoolList
-from ..common_primitives import (
+from ... import core
+from ..autograd_common_primitives import (
     add,
     buffer,
     cartesian_diff,
@@ -71,6 +68,7 @@ from ..common_primitives import (
     tuple_converter,
     union,
 )
+from ..utils import NestedFloatOrIntOrBoolList, find_dominant_type, is_tuple_int
 from .utils import (
     broadcast_to_highest,
     calc_prob_matrix,
